backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.0
python-dotenv==1.0.1
python-multipart==0.0.9
requests==2.31.0
xmltodict==0.13.0
google-generativeai>=0.3.0
langchain>=0.0.350
langchain-community>=0.0.10
langchain-google-genai>=0.0.5
sentence-transformers==2.6.1
faiss-cpu==1.8.0.post1
PyPDF2==3.0.1